requirements.txt
musicbrainzngs
psycopg2-binary
clickhouse-connect
